Return data is now being stored in the Redux store.
