+++ src/App.css
@@ -1,32 +1,3 @@
-header {
-  background-color: blue;
-  color: #ffffff;
-  font-weight: bold;
-  height: 45px;
-  vertical-align: middle;
-  line-height: 45px;
-}
-
 .App {
   height: 100%;
-}
-
-.App-Body {
-  margin: 10px;
-}
-
-.Body-Top {
-  display: grid;
-  grid-gap: 10px;
-  grid-template-columns: auto 100px;
-}
-
-.Body-Top .ActionButton {
-  width: 100%;
-  font-size: 1.2em;
-}
-
-.Body-Top-Button {
-  display: flex;
-  align-items: center;
 }
