+++ src/actions/actionTypes.js
@@ -5,5 +5,4 @@ These are just Unique string representations of all the unique actions.
 
 export const SET_URL = 'SET_URL';
 export const SET_QUERY = 'SET_QUERY';
-export const FETCH_STUFF = 'FETCH_STUFF';
-export const RECEIVE_STUFF = 'RECEIVE_STUFF';
+export const SET_RECEIVED_DATA = 'SET_RECEIVED_DATA';
