+++ src/components/SplitQuery.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import './SplitQuery.css';
 import {Query}  from "../components";
+import {connect} from 'react-redux';
 
 class SplitQuery extends Component {
     render() {
@@ -15,10 +16,12 @@ class SplitQuery extends Component {
             <div className="Split">
                 <div className="SplitGrid">
                     <div className="SplitItem">
-                        <Query query={this.props.query1} onClick={() => queryClick(1)}/>
+                        <Query query={this.props.query1}
+                               onClick={() => queryClick(1)}/>
                     </div>
                     <div className={secondItem}>
-                        <Query query={this.props.query2} onClick={() => queryClick(2)}/>
+                        <Query query={this.props.query2}
+                               onClick={() => queryClick(2)}/>
                     </div>
                 </div>
             </div>
@@ -26,4 +29,12 @@ class SplitQuery extends Component {
     }
 }
 
-export default SplitQuery;
+const mapStateToProps = state => {
+    return {
+        query1: state.query.query1,
+        query2: state.query.query2,
+    };
+};
+
+
+export default connect(mapStateToProps)(SplitQuery);
