+++ src/components/TopLevel.css
@@ -1,0 +1,32 @@
+header {
+  background-color: blue;
+  color: #ffffff;
+  font-weight: bold;
+  height: 45px;
+  vertical-align: middle;
+  line-height: 45px;
+}
+
+.TopLevel {
+  height: 100%;
+}
+
+.TopLevel-Body {
+  margin: 10px;
+}
+
+.Body-Top {
+  display: grid;
+  grid-gap: 10px;
+  grid-template-columns: auto 100px;
+}
+
+.Body-Top .ActionButton {
+  width: 100%;
+  font-size: 1.2em;
+}
+
+.Body-Top-Button {
+  display: flex;
+  align-items: center;
+}
