+++ src/components/index.js
@@ -12,6 +12,7 @@ import Menu from "./Menu";
 import MenuIcon from "./MenuIcon";
 import Split from "./Split";
 import ActionButton from "./ActionButton";
+import TopLevel from "./TopLevel";
 
 
 export {Endpoint};
@@ -28,3 +29,4 @@ export {Query};
 export {ResultRow};
 export {ResultTable};
 export {KeySelect};
+export {TopLevel};
